refactor: split management HTTP policy helpers
